Separated the components in various files just for being more dry and effective
