+++ src/components/Quote.js
@@ -1,0 +1,22 @@
+import {
+  BrowserRouter,
+  Routes,
+  Route,
+  NavLink,
+} from 'react-router-dom';
+import styleQuote from './Quote.css';
+
+const Quote = () => {
+  return (
+    <div className="quote_container" style={styleQuote}>
+      <h2>
+        Mathematics is not about numbers, equations, computations, or
+        algorithms: it is about understanding.
+        <br />
+        -William Paul Thurston
+      </h2>
+    </div>
+  );
+};
+
+export default Quote;
